Add total weight collected display

+++ Assets/Scripts/Product.cs
@@ -8,7 +8,7 @@ public class Product : Interactable
 {
     [Header("Product")]
     public string productName;
-    public float weight;
+    public int weight;
 
     public const float buyDuration = 1.5f;
     public const float moveY = 0.5f;
